requirements.txt
google-generativeai==0.8.3
yfinance==0.2.49